refactor to separate spec from builder
