+++ src/xdmfSpecification.hpp
@@ -1,0 +1,60 @@
+#ifndef PETSCXDMFGENERATOR_XDMFSPECIFICATION_H
+#define PETSCXDMFGENERATOR_XDMFSPECIFICATION_H
+
+#include<string>
+#include<vector>
+#include<memory>
+#include "hdfObject.hpp"
+
+namespace petscXdmfGenerator {
+enum FieldLocation {NODE, CELL};
+enum FieldType {SCALAR, VECTOR, TENSOR, MATRIX};
+
+class XdmfSpecification {
+   private:
+    struct Description {
+        std::string path = "";
+        unsigned long long number = 0;
+        unsigned long long numberCorners = 0;
+        unsigned long long dimension = 0;
+    };
+
+    struct FieldDescription {
+        std::string name = "";
+        std::string path = "";
+        std::vector<unsigned long long> shape;
+        FieldLocation fieldLocation;
+        FieldType fieldType;
+    };
+
+//    struct GridDescription{
+        std::string name = "domain";
+        // store each type of geometry/topology data
+        Description topology;
+        Description hybridTopology;
+        Description geometry;
+
+        // store the field data
+        std::vector<FieldDescription> fields;
+
+        std::vector<double> time;
+//    };
+
+    // Store the path to the file
+    std::string hdf5File;
+
+    // helper functions
+    static void GenerateFieldsFromPetsc(std::vector<FieldDescription>& fields, const std::vector<std::shared_ptr<petscXdmfGenerator::HdfObject>>& hdfFields,FieldLocation location);
+    static std::shared_ptr<petscXdmfGenerator::HdfObject> FindPetscHdfChild(std::shared_ptr<petscXdmfGenerator::HdfObject>& root, std::string name);
+    // Allow the builder to access
+    friend class XdmfBuilder;
+
+   public:
+    // provide generator functions
+    static std::shared_ptr<XdmfSpecification> FromPetscHdf(std::shared_ptr<petscXdmfGenerator::HdfObject>);
+
+
+};
+
+}
+#endif  // PETSCXDMFGENERATOR_XDMFSPECIFICATION_H
