+++ tests/integrationTests.cpp
@@ -19,8 +19,8 @@ class PETScHdf5ToXdmfTestFixture : public ::testing::TestWithParam<std::string> 
 TEST_P(PETScHdf5ToXdmfTestFixture, ShouldGenerateExpectedXml) {
     // arrange
     auto hdfObject = std::make_shared<petscXdmfGenerator::HdfObject>(inputFilePath);
-    auto builder = petscXdmfGenerator::XdmfBuilder::FromPetscHdf(hdfObject);
-
+    auto specification = petscXdmfGenerator::XdmfSpecification::FromPetscHdf(hdfObject);
+    auto builder = petscXdmfGenerator::XdmfBuilder(specification);
     // read in the expectedResults
     std::ifstream expectedResultFile(expectedOutputFilePath);
     std::stringstream expectedOutput;
@@ -29,7 +29,7 @@ TEST_P(PETScHdf5ToXdmfTestFixture, ShouldGenerateExpectedXml) {
     std::stringstream resultStream;
 
     // act
-    auto xml = builder->Build();
+    auto xml = builder.Build();
     xml->PrettyPrint(resultStream);
 
     // assert
